Scale rate value to volume and/or mass

This is a conversion function. It can convert a dimensionless unit of rate, derived from calc_rate into volume-adjusted (e.g to the container), VO_2 or mass specific (i.e. to the specimen), MO_2 rate.

convert_rate(x, o2.unit = NULL, time.unit = NULL, output.unit = NULL,
  volume = NULL, mass = NULL, S = 35, t = 25, P = 1.013253)

Arguments

x
numeric, or objects of class calc_rate(), auto_rate() or adjust_rate().

o2.unit
string. The dissolved oxygen unit to convert from. Check unit_args().

time.unit
string. The time unit to convert from. Check unit_args().

output.unit
string. The output unit to convert to. Check unit_args().

volume
numeric. Volume in litres. This is the volume of the respirometry chamber, not the specimen(s).

mass
numeric. Mass/weight in kg. This is the mass of the specimen if you wish to calculate mass-specific rates.

S
numeric. Salinity. Defaults to 35.

t
numeric. Temperature. Defaults to 25 (degress celcius).

P
numeric. Air pressure. Defaults to 1.013253 (bar).

Value

A list object.

Details

The function uses an inernal database and a fuzzy string matching algorithm to accept various unit formatting styles.

For example, 'mg/l', 'mg/L', 'mgL-1', 'mg l-1', 'mg.l-1' are all the same. Use unit_args() to view a list of usable unit strings.

Examples

convert_rate(7.5, o2.unit = 'mg/l', time.unit = 's',
  output.unit = 'mg/min/kg', volume = 1.2, mass = 0.5)

#> Input:
#> [1] 7.5
#> [1] "mg/L" "sec" 
#> Output:
#> [1] 1080
#> [1] "mg/min/kg"