Multi-column respirometry data of the sea urchin, Heliocidaris Erythrogramma, including background respiration

Oxygen consumption data of 16 individual Heliocidaris erythrogramma specimens. In addition, there are 2 measurements of background respiration. These were trial runs to determine the precision of the new Vernier Optical DO probes (ODO-BTA), and thus chamber volumes and specimen masses were not recorded.

urchins.rd

Format

A data frame object consisting of one column of time, 16 columns of urchin oxygen consumption (a-p) and 2 columns of background oxygen consumption (b1 & b2). There are 271 rows of data spanning 45 minutes.

Details

Dissolved oxygen units: mg/L

Time units: minutes

Chamber volumes (L): N/A

Specimen masses (kg): N/A

Author

Januar Harianto