Calculate flowthrough rate of change in dissolved oxygen

Calculates rate of $O_2$ uptake in flowthrough respirometry given either:

calc_rate.ft(df = NULL, from = NULL, to = NULL, by = "row",
  inflow.col = NULL, outflow.col = NULL, inflow.o2 = NULL,
  outflow.o2 = NULL, delta.o2 = NULL, flowrate = NULL, plot = TRUE)

Arguments

df
data frame. Defaults to NULL. Will process a data frame if it is provided here.

from
numeric. Defaults to NULL. Defines the upper bound of the data frame to subset. Subsetting is based on the argument: by.

to
numeric. Defaults to NULL. Defines the lower bound of the data frame to subset. Subsetting is based on the argument: by.

by
character. Defaults to "row".

inflow.col
numeric. Defaults to NULL. This selects the incurrent column if a data frame ('df') is provided.

outflow.col
numeric. Defaults to NULL. This selects the outcurrent column if a data frame ('df') is provided.

inflow.o2
numeric. Single value or vector. Incurrent oxygen concentration. If outflow.o2 is a vector, this must be a single value or a vector of equal length.

outflow.o2
numeric. Single value or vector. Excurrent oxygen concentration.

delta.o2
numeric. Single value or vector. An O2 differential between inflow and outflow $O_2$. Used only if inflow.o2 and outflow.o2 are NULL.

flowrate
numeric vector. The flow rate. No unit of measurement is expected; you will specify it when you perform conversions later on.

plot
logical. Defaults to TRUE. Plots the data.

Value

An object of class "calc_rate.ft".

Details

A flow-rate and both inflow and outflow oxygen concentrations 2. A flow-rate and an $O_2$ differential between inflow and outflow concentrations

Can return a single value, or multiple and mean values based on continuous data.

Examples

NULL

#> NULL