Import respirometry system output files

Automatically read data from different respirometry hardware and software systems. The aim is to work with most commercial DO sensors available in the market with little input from the user. The function extracts data columns from the file and generally cleans up column names to make the data easier to work with.

import_file(path, export = FALSE)

Arguments

path
string. Path to file.

export
logical. If TRUE, exports the data to the same directory, determined by the path parameter above.

Value

a data frame object of all columned data

Details

Currently works for:

Firesting Logger

Pyro Oxygen Logger (another name for Firesting)

PreSens OXY10

PreSens OXY4

PreSens (OxyView generic, including multiplate systems)

PreSens/Loligo 24-Well Multiplate System (Excel files)

MiniDOT

Loligo AutoResp ('_raw' files output, not metadata files)

Loligo Witrox (same as AutoResp, without metadata)

Vernier (raw qmbl, csv, txt, (not yet gmbl))

NeoFox

Qbox Aqua

We are always looking for sample files to improve the function. Please send them to us via email, or via a Github issue.

While the devices listed above are supported, the import functionality may not be complete due to limited access to output files. This will improve over time as users provide feedback. We are releasing this as it is, without any warranty, so that some people can still benefit from the functionality as it gets better. Users should be careful using this function, and be prepared to to import data by themselves since it is a fundamental skill in R.

Examples

NULL

#> NULL