Calculate critical oxygen tension, \(P_{crit}\)

Critical oxygen tension can be calculated either by the "broken-stick" regression (BSR) approach, adopted from Yeager and Ultsch (1989), or the segmented regression approach, presented by Muggeo (2003).

pcrit(df, width = floor(0.1 * nrow(df)), has.rate = FALSE, plot = TRUE,
  parallel = TRUE)

Arguments

df
data frame or object of class inspect_data. This is the data to analyse.

width
numeric. Defaults to floor(0.1*nrow(df)).

has.rate
logical. Defaults to FALSE.

plot
logical. Defaults to TRUE.

parallel
logical. Defaults to TRUE. Should parallel processing be used?

Value

A list object of class pcrit.

Details

To calculate Pcrit, it is necessary to provide data in the form of the rate of change in oxygen (RO2/MO2) against dissolved oxygen (DO) concentration. If you have a single dataset that only measured DO over time, pcrit() can still automatically determine RO2/MO2 against DO by performing a rolling regression of the data and match it agains the rolling mean of the DO data. This is also the default method. If you already have calculated RO2/MO2 elsewhere, the argument has.rate = "TRUE" must be used. The width of the rolling regression is determined by the width argument. In most cases, the default width (10

References

Yeager DP, Ultsch GR (1989) Physiological regulation and conformation: A BASIC program for the determination of critical points. Physiological Zoology 62:888–907. doi: 10.1086/physzool.62.4.30157935

Muggeo V (2008) Segmented: an R package to fit regression models with broken-line relationships. R News 8:20–25.

Examples

# The example below should work (uncomment to run)
# pcrit(squid.rd)