Print examples of unit arguments for use in convert_DO() and convert_rate()

This is a simple function, requiring no argument. It shows available units that can be used in the functions convert_DO() and convert_rate().

unit_args()

Value

A list.

Details

Available units for convert_DO():

"mg/L", "ug/L", "mmol/L", "umol/L", "mL/L", "mg/kg", "ug/kg", "mmol/kg", "umol/kg", "mL/kg", "

Available units for convert_rate(): O2

"mg/L", "ug/L", "mmol/L", "umol/L", "mL/L", "mg/kg", "ug/kg", "mmol/kg", "umol/kg", "mL/kg", "Torr", "hPa", "kPa", "mmHg", "inHg" time

"s", "m", "h" mass

"ug", "mg", "g", "kg"

Examples

unit_args()

#> Note: A string-matchin algorithm is used to identify units. e.g. all of these are the same: mg/L; mg/l, mg L-1, mgL-1, mg per litre, mg.l-1, mg.L-1
#> 
#> O2 Units
#>  [1] "mg/L"    "ug/L"    "mmol/L"  "umol/L"  "mL/L"    "mg/kg"   "ug/kg"  
#>  [8] "mmol/kg" "umol/kg" "mL/kg"   "%"       "Torr"    "hPa"     "kPa"    
#> [15] "mmHg"    "inHg"   
#> 
#> Time units
#> [1] "s" "m" "h"
#> 
#> Mass units
#> [1] "ug" "mg" "g"  "kg"