Print examples of unit arguments for use in convert_DO() and convert_rate()

This is a simple function, requiring no argument. It shows available units that can be used in the functions convert_DO() and convert_rate(). Some units also require temperature (t), salinity (S), and atmospheric pressure (P) to be specified.

unit_args()

Value

A list.

Details

Available units for convert_DO():

Do NOT require t, S and P:

"mg/L", "ug/L", "mmol/L", "umol/L"

Require t, S and P:

"mL/L", "mg/kg", "ug/kg", "mmol/kg", "umol/kg", "mL/kg", "%" (N.B. = % Air Saturation), "Torr", "hPa", "kPa", "mmHg", "inHg"

Available units for convert_rate():

O2

Do NOT require t, S and P:

"mg/L", "ug/L", "mmol/L", "umol/L"

Require t, S and P:

"mL/L", "mg/kg", "ug/kg", "mmol/kg", "umol/kg", "mL/kg", "%" (N.B. = % Air Saturation), "Torr", "hPa", "kPa", "mmHg", "inHg"

Time

"s", "m", "h"

Output mass

"ug", "mg", "g", "kg"

Examples

unit_args()

#> Note: A string-matching algorithm is used to identify units. 
#> E.g. all of these are the same: mg/L; mg/l, mg L-1, mgL-1, mg per litre, mg.l-1, mg.L-1
#> 
#> O2 Units - Do not require t, S and P
#> [1] "mg/L"   "ug/L"   "mmol/L" "umol/L"
#> 
#> O2 Units - Require t, S and P
#>  [1] "mL/L"    "mg/kg"   "ug/kg"   "mmol/kg" "umol/kg" "mL/kg"   "%"      
#>  [8] "Torr"    "hPa"     "kPa"     "mmHg"    "inHg"   
#> 
#> Time units
#> [1] "s" "m" "h"
#> 
#> Output mass units
#> [1] "ug" "mg" "g"  "kg"