Print examples of unit arguments for use in convert_DO() and convert_rate()

This is a simple function, requiring no argument. It shows available units that can be used in the functions convert_DO() and convert_rate(). Some units also require temperature (t), salinity (S), and atmospheric pressure (P) to be specified.

unit_args()

Value

A list.

Details

Available units for convert_DO():

Do NOT require t, S and P:

"mg/L", "ug/L", "mmol/L", "umol/L"

Require t, S and P:

"mL/L", "mg/kg", "ug/kg", "mmol/kg", "umol/kg", "mL/kg", "\ "mmHg", "inHg"

Available units for convert_rate():

O2

Do NOT require t, S and P:

"mg/L", "ug/L", "mmol/L", "umol/L"

Require t, S and P:

"mL/L", "mg/kg", "ug/kg", "mmol/kg", "umol/kg", "mL/kg", "\ "mmHg", "inHg"

Time

"s", "m", "h"

Output mass

"ug", "mg", "g", "kg"

Examples

unit_args()

#> Note: A string-matching algorithm is used to identify units. 
#> Example 1: These are recognised as the same: 'mg/L', 'mg/l', 'mg L-1', 'mg per litre', 'mg.L-1'
#> Example 2: These are recognised as the same: 'hour', 'hr', 'h'
#> 
#> # Input Units # --------------------------------------
#> Oxygen concentration units should use SI units (`L` or `kg`) for the denominator.
#> 
#> Oxygen Concentration or Pressure Units - Do not require t, S and P
#> [1] "mg/L"   "ug/L"   "mmol/L" "umol/L"
#> Oxygen Concentration or Pressure Units - Require t, S and P
#>  [1] "mL/L"    "mg/kg"   "ug/kg"   "mmol/kg" "umol/kg" "mL/kg"   "%Air"   
#>  [8] "%O2"     "Torr"    "hPa"     "kPa"     "mmHg"    "inHg"   
#> 
#> Volume units for use in flow rates in calc_rate.ft and convert_rate.ft
#> (e.g. as in 'ml/min', 'L/s', etc.)
#> [1] "uL" "mL" "L" 
#> 
#> Time units (for 'time.unit' or as part of 'flowrate.unit')
#> [1] "sec"  "min"  "hour" "day" 
#> 
#> Mass units
#> [1] "ug" "mg" "g"  "kg"
#> 
#> Area units
#> [1] "mm2" "cm2" "m2"  "km2"
#> 
#> # Output Units in 'convert_rate' & 'convert_rate.ft' #
#> Must be in correct order, with no special characters other than the separator:
#> 
#> Absolute rates:        O2/Time       e.g. 'mg/sec',     'umol/min',     'mL/h'
#> Mass-specific rates:   O2/Time/Mass  e.g. 'mg/sec/ug',  'umol/min/g',   'mL/h/kg'
#> Area-specific rates:   O2/Time/Area  e.g. 'mg/sec/mm2', 'umol/min/cm2', 'mL/h/m2'
#> 
#> Output Oxygen amount units
#> [1] "ug"   "mg"   "umol" "mmol" "mol"  "mL"  
#> 
#> Output Time units
#> [1] "sec"  "min"  "hour" "day" 
#> 
#> Output Mass units for mass-specific rates
#> [1] "ug" "mg" "g"  "kg"
#> 
#> Output Area units for surface area-specific rates
#> [1] "mm2" "cm2" "m2"  "km2"